Add Graphql API and implement user signup

+++ database/migrations/2024_11_25_210112_add_api_token_to_users_table.php
@@ -15,8 +15,12 @@ return new class extends Migration
 
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('api_token');
-        });
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('api_token');
+            });
+        } else {
+            // Handle SQLite specific logic, like using the workaround
+        }
     }
 };
